requirements.txt
python-magic-bin==0.4.14
Pillow==11.0.0
exifread==3.0.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==0.4.0
odfpy==1.4.1
PyMuPDF==1.24.12
python-magic==0.4.27